+++ components/story.js
@@ -1,0 +1,5 @@
+const story = document.getElementById('story')
+
+export const getPaper = () => {
+    const paper = document.createElement('div')
+}
