+++ css/output.css
@@ -1,3 +1,12 @@
+body::-webkit-scrollbar {
+  background-color: #222;
+  width: 7px;
+}
+body::-webkit-scrollbar-thumb {
+  background-color: #666;
+  width: 15px;
+}
+
 * {
   padding: 0;
   margin: 0;
@@ -9,6 +18,14 @@ body {
 }
 
 /* -- */ /* CLASES */ /* -- */
+#buildSection, #horrors, #news, #story {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
 .class {
   display: flex;
   flex-direction: column;
@@ -65,13 +82,14 @@ body {
   display: none;
   align-items: center;
   justify-content: center;
-  position: absolute;
+  position: fixed;
   width: 100%;
-  height: 90%;
+  height: 100vh;
+  top: 0;
   background-color: #222;
   grid-template-areas: "gridL imgClass gridR buildDesc" "btns btns btns btns";
+  z-index: 1030;
 }
-
 .build .buildGridL {
   grid-area: gridL;
 }
@@ -116,12 +134,7 @@ body {
   border: solid 2px #000;
   justify-self: center;
   align-self: center;
-}
-
-.green.bg {
-  background-color: transparent;
-  border: none;
-  border: solid 2px transparent;
+  margin: 5px 0;
 }
 
 .green.bg:hover {
@@ -147,7 +160,6 @@ body {
   color: #fff;
   cursor: pointer;
   border-radius: 20px;
-  z-index: 5;
 }
 
 .build .close.closeModal {
@@ -198,12 +210,14 @@ body {
 /* -- */ /* -- // FIN BUILD // -- */ /* -- */
 /* -- */ /* -- // BUILD DESCRIPTION // -- */ /* -- */
 .bgModal {
-  width: 100%;
-  height: 150%;
-  top: -200px;
-  left: 0;
-  position: absolute;
+  width: 100vw;
+  height: 150vh;
+  position: fixed;
+  display: flex;
+  justify-content: center;
+  align-items: center;
   background-color: hsla(0, 9%, 9%, 0.5);
+  z-index: 1020;
 }
 
 .buildDescription {
@@ -226,17 +240,13 @@ body {
 
 .skillsModal {
   max-height: 500px;
-  min-width: 200px;
+  width: 70%;
   padding: 5px;
   overflow-y: scroll;
   position: absolute;
   text-align: left;
   background-color: #444;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
 }
-
 .skillsModal img {
   width: 50px;
 }
@@ -263,11 +273,12 @@ body {
     justify-content: center;
     height: auto;
     width: 100%;
+    height: 100%;
     gap: 10px;
     margin-top: 10px;
   }
   .build .buildGridR img, .build .buildGridL img {
-    width: 110px;
+    width: 80px;
   }
   .build .skillsModal {
     width: 90%;
@@ -296,13 +307,14 @@ body {
 }
 @media (max-width: 630px) {
   .build {
-    grid-template-areas: "imgClass imgClass" "buildDesc buildDesc" "gridR gridR" "gridL gridL" "btns btns";
+    grid-template-areas: "imgClass _" "buildDesc buildDesc" "gridR gridR" "gridL gridL" "btns btns";
   }
   .build .imgContainer .classImg {
     display: none;
   }
   .build .imgContainer .charm {
     width: 85px;
+    margin: 10px 0;
     position: relative;
     left: 0;
   }
@@ -311,6 +323,7 @@ body {
   }
   .build .buildGridR, .build .buildGridL {
     overflow-x: scroll;
+    overflow-y: hidden;
     justify-content: flex-start;
   }
   .build .btnBuild {
@@ -337,7 +350,6 @@ body {
 /* -- */ /* -- // FIN BUILD DESCRIPTION // -- */ /* -- */
 /* boss */
 .mainBoss {
-  background-image: url(../img/fondoseccion_boss.jpg);
   background-position: center;
   height: 100%;
   width: 100%;
@@ -458,12 +470,10 @@ nav a:hover {
   color: rgb(108, 117, 125) !important;
   transition: color 0.4s ease-in-out;
 }
-
 nav img:hover {
   filter: brightness(150%);
   transition: filter 0.4s ease-in-out;
 }
-
 
 .blur {
   content: "";
@@ -472,14 +482,12 @@ nav img:hover {
   background: linear-gradient(to right, rgb(0, 0, 0), transparent);
 }
 
-
 @media (max-width: 1199.98px) {
   #inicio {
     align-items: center;
   }
   .inicio-video {
     width: 100% !important;
-
     filter: brightness(40%);
   }
   nav {
@@ -497,9 +505,7 @@ nav img:hover {
   }
   .text-title {
     width: 75% !important;
-
     align-self: center;
-
   }
   .blur {
     left: 0;
@@ -508,7 +514,9 @@ nav img:hover {
     bottom: 0;
     background: linear-gradient(to top, rgb(0, 0, 0), transparent);
   }
-
+}
+.paper {
+  background-image: url("https://previews.123rf.com/images/roystudio/roystudio1202/roystudio120200165/12658204-hoja-de-papel-con-delicado-patr%C3%B3n-para-utilizar-como-fondo.jpg");
 }
 
 /*# sourceMappingURL=output.css.map */
