+++ db/story.json
@@ -1,0 +1,8 @@
+[
+    {
+        "page": 1
+    },
+    {
+        "page": 2
+    }
+]
