+++ sass/_boss.scss
@@ -1,7 +1,6 @@
 /* boss */
 
 .mainBoss {
-    background-image: url(../img/fondoseccion_boss.jpg);
     background-position: center;
     height: 100%;
     width: 100%;
