+++ sass/_general.scss
@@ -1,0 +1,10 @@
+body {
+  &::-webkit-scrollbar {
+    background-color: #222;
+    width: 7px;
+  }
+  &::-webkit-scrollbar-thumb {
+    background-color: #666;
+    width: 15px;
+  }
+}
