+++ sass/_story.scss
@@ -1,23 +1,3 @@
-
-
-
-
-#story {
-    background: url("../img/story" )center no-repeat ;
-    background-size: contain;
-
-
-    height: 85vh;
-
-
-
-
-
-
-
-
-
-
-
-
+.paper {
+    background-image: url('https://previews.123rf.com/images/roystudio/roystudio1202/roystudio120200165/12658204-hoja-de-papel-con-delicado-patr%C3%B3n-para-utilizar-como-fondo.jpg');
 }
