+++ sass/input.scss
@@ -1,3 +1,4 @@
+@use "general";
 @use "build";
 @use "boss";
 @use "fonts";
